Remove quictls
The quictls/openssl repo is no longer actively maintained and security
issues already start piling up, as is to be expected since it contains
most of the careless and poorly reviewed churn that happened between
OpenSSL 1.1 and now.

We can reinstate quictls once there is a release of quictls that keeps
up with the security fixes from their upstream.

https://github.com/quictls/quictls/issues/244