split up the CMake port, providing binaries and the associated files required for building in "cmake-core" (subpackage of the "core" subport), helptext files in "cmake-help" (different subpackage of "core", to reduce churn of ~2100 files during bulk builds), and manual pages in "cmake-man" (different subport).
Upstream does not distribute manpages in a ready to use format, instead
requiring the Python-based Sphinx tool to build them. Splitting these off
to a separate subport avoids the need to build dozens of packages before
CMake-based ports can be reached in a bulk build.

A meta-package is provided so that "pkg_add cmake" installs core files
plus manuals as expected.

ok rsadowski

module and dependent ports will be adjusted in subsequent commit