Update KiCAD to 8.0.9
This update fixes the current py3.12 issue and unbreaks it. Unfortunately, this
update introduces a new problem: footprint handling segmentation fault.

This can result in projects breaking when they are loaded with footprints. For
this reason, the port remains as BROKEN.

Currently there is a workaround to use KiCad 8 under OpenBSD.
Base libcxxabi must be built with _LIBCXX_DYNAMIC_FALLBACK.

You can find instructions and more information here:
https://marc.info/?l=openbsd-ports&m=172511950120387&w=2