Update KiCAD to 9.0.0
It looks like upstream has fixed some things in their shared+static libs as
mentioned in e32b26eb [2]. As a result, I no longer see the error reported in
[1]. This probably means that we don't need the
"_LIBCXXABI_FORGIVING_DYNAMIC_CAST" om out libcxxabi.

To be on the safe side, I need a real word user test. For this reason I have
marked KiCAD 9.0 as unbroken. (also on powerpc, the description points to the
actual error).

Thanks to espie and jca@ for dealing with this nasty s***.

1: https://gitlab.com/kicad/code/kicad/-/issues/18621
2: https://gitlab.com/kicad/code/kicad/-/commit/e32b26ebeb300ef5c6edc773d6852c60b85f980d