Import devel/py-test-run-parallel 0.4.0.
This pytest plugin takes a set of tests that would be normally be run serially
and execute them in parallel.

The main goal of pytest-run-parallel is to discover thread-safety issues that
could exist when using C libraries, this is of vital importance after PEP703,
which provides a path for a CPython implementation without depending on the
Global Interpreter Lock (GIL), thus allowing for proper parallelism in programs
that make use of the CPython interpreter.

adds the pytest 'thread_unsafe' marker, needed to run tests for pydantic update

ok tb@