sysutils/tmate: compatibility with new imsg API
We had the imsg earthquake which includes renaming of imsg_init into
imsgbuf_init, which leads a configure to a conclusion that tmate needs
compat/imsg.c and compat/imsg-buffer.c where additional socket() is
used. That ends to a need extention of pledge().

This was a wrong way, the right one is support the new imsg API.

OK: tb@