xmms2: mark broken
This doesn't build with pkgconf due to linking issues with vorbis:

ld: error: undefined symbol: vorbis_info_init
>>> referenced by ripper.c:181 (../src/clients/vistest/ripper.c:181)
>>>               src/clients/vistest/ripper.c.3.o:(main)

etc.

Maybe a brave soul will find the time and courage wade into the custom
WAF mess and figure out what causes this or at how to update this port
to its latest and greatest release 0.9.5, which no one felt the need to
do in the last 19 REVISIONS...

Meanwhile we want to move forward with pkgconf.