update to got 0.112 and make the update-plist target happy
op@ pointed out that update-plist was broken and sent a fix. Reuse parts
of that fix to create /var/www/got/public in the fake install area, and
add a @sample directive as suggested by sthen@. The @sample causes
update-plist to accept the directory as it appears in the packing list.

Upstream changelog:

- remove /tmp/got-importmsg temp files when import commit message is left empty
- rely on secondary _gotwebd groups if repos_path is not owned by _gotwebd group
- fix unrelated errors being reported if a histedit operation is aborted
- implement support for protected references in gotsys.conf and gotsysd
- plug memory leaks in some libexec helpers and in the gitconfig parser
- stop needlessly opening the repository whenever a work tree is opened