archivers/zstd: workaround for missing LESSOPEN support in zstdless
The `zstdless` script, which is part of the zstd package, relies on the
LESSOPEN mechanism, which was removed from OpenBSD over a year ago. As a
result, `zstdless` is currently broken. Work around this by replacing
`zstdless` with a script using `zstdcat` and `less`.

Issue reported and workaround suggested by Tom Szilagyi <tom DOT
szilagyi AT altmail DOT se>.