update to isc-bind-9.20.9
https://kb.isc.org/docs/cve-2025-40775
"When an incoming DNS protocol message includes a Transaction Signature
(TSIG), BIND always checks it. If the TSIG contains an invalid value in
the algorithm field, BIND immediately aborts with an assertion failure.

Impact: Denial-of-service. By sending specific messages to the server,
an attacker can cause named to terminate unexpectedly.

Authoritative servers are affected by this vulnerability.

Resolvers are affected by this vulnerability."