stop bundling blender's embedded copy of Python
tb@ reported that blender still breaks at build time. The latest breakage
was something to do with missing chardet.

As a quick fix let's stop bunding blender's copy of Python and see if that
minimizes fallout in bulks.

Folks interested in blender can continue to explore if there are better
approaches.

ok tb@