update to clamav-1.4.3
CVE-2025-20260: Fixed a possible buffer overflow write bug in the PDF
file parser that could cause a denial-of-service (DoS) condition or
enable remote code execution.

CVE-2025-20234: Fixed a possible buffer overflow read bug in the UDF
file parser that may write to a temp file and thus disclose information,
or it may crash and cause a denial-of-service (DoS) condition.

Fixed a possible use-after-free bug in the Xz decompression module in
the bundled lzma-sdk library.