Update Plasma 6.4
The most parts are straightforward as usual but in 6.4 the KDE
Kwin team split kwin into kwin-x11 and kwin (wayland). This seems
to be the sign that X11 is no longer of interest and we are
focussing on Wayland.

As we currently only support X11, kwin-x11 has been added as a runtime
dependency to kwin. So nobody should have to install anything later.

This ports update also includes Aurorae; a theme engine for KWin window
decorations.

https://kde.org/announcements/plasma/6/6.4.0/

This update also includes Bugfix Release for June and July.

Thanks yaydn for feedback and testing and kn@ for review