octave: apply usual fix for build with opaque __sFILE:
Patch gnulib/fseeko.c.  gnulib is doing this to make fflush() behaivour
to match what they expect.  It accesses internal of __sFILE to change
behavior of fseeko(), but we are hiding internal of __sFILE and our
fseeko() is compatible with what gnulib expects already.  Then the
patch is to just use our fseeko().