hard-code the W^X code path for PolyML on OpenBSD
The code in PolyML tries to detect whether W^X is enabled or not at
runtime. On OpenBSD switching behaviour at run-time is not so straight
forward, so just short-circuit the test and hard-code to use the W^X
code path.

Discussion with sthen@ was triggered by kili@ (who sets kern.wxabort=1 and
ran into the run-time detection failing at build time).

The patch is not strictly needed, but we will apply it locally for now.

ok sthen@, kili@