update to anubis-1.21.0
This has some fairly big changes, see
https://github.com/TecharoHQ/anubis/releases/tag/v1.21.0

Challenge creation to was redone to fix the '"invalid response" after
"success"' bug. Challenges now use a per-request random number rather
than client metadata (which Anubis was hoping would be consistent
between requests, but it turned out not to be the case). During the
upgrade to v1.21.0, and when v1.21.0/newer is restarted if using the
(default) in-memory storage backend, you may see a higher rate of failed
challenges than normal. If this behaviour around restarts is a problem
for you, look into configuring Anubis to use persistent storage.

Unveil patch dropped, if someone wants this back it will need reworking
to cope with the configurable path for on-disk storage.