Import sysutils/ansible-runner 2.4.1, from Mikolaj Kucharski
Ansible Runner is a tool and Python library that helps when
interfacing with Ansible directly or as part of another system whether
that be through a container image interface, as a standalone tool, or
as a Python module that can be imported. The goal is to provide a
stable and consistent interface abstraction to Ansible.

ok caspar@