py-qt6: split configure and build steps
The --no-make directive of sip-build lets 'make configure' stop before
the actual make command is invoked, so the build step can work as usual.

Pass MAKE_JOBS to the MAKE_FLAGS to allow parallelizing the build. This
way the configure step is "only" a few minutes rather than hours and the
parallelized build step can be done in a fraction of an hour on modern
hardware, opening this up to DPB_PROPERTIES = parallel if we so choose.

tests/ok landry sthen
Yes, please!!!!!! phessler