update to latest git commit c27ebe5
This release adds support for PIC/PIE and has merged the handling of
unknown warnings. With this commit, we are finally able to compile
bin/cat on OpenBSD with no modifications:

	$ env CC=ccomp PIPE= make

And then both "cat cat.c" and "cat -n cat.c" work (the latter is a useful
test on amd86/IBT as it covers the case of code with a jump table).