Split ksh93 into -main and -libshell subpackages.
The ksh93 binary in the -main package has libshell etc. linked in statically,
while -libshell provides the libraries and corresponding headers.

This is done mainly to prevent a problem in bulk builds because of libshell's
@conflict with devel/libast, where x11/cde will depend on ksh93, and eterm
depends on devel/libast.  It also makes the package a bit slimmer for people
who don't need libshell.

prompted by tb@, ok aja@.