Import ruby34-openfact-5.1.0;  Input OK jeremy
This is the only new dependency for upcoming OpenVox, a Puppet fork.

OpenFact is a community implementation of Facter, a command-line tool that
gathers basic facts about nodes (systems) such as hardware details, network
settings, OS type and version, and more. These facts are made available as
variables in your Puppet manifests and can be used to inform conditional
expressions in Puppet.