Fetch a distfile of the recently tagged librsvg-gtk-2.40.23 and use it on !rust (i.e. old-librsvg) archs to unbreak the build.  ok aja@
gtk+4 started depending on librsvg - it prefers 2.48+ (which requires
rust) but is willing to use 2.40.23+ as a fallback (and has the build
infrastructure to handle a local statically-linked build). However,
there's no standalone release of 2.40.23 yet, only a git tag, and
creating the bootstrap files (automake etc) from this didn't go
smoothly, so this seems a sane approach for now.

https://gitlab.gnome.org/GNOME/gtk/-/commit/afa0d2341a442d356f91ee0a9f7b168cf14d628a