Update to pinentry-1.3.2
Add missing dep on desktop-file-utils for -qt5 and -qt6.
Also the -qt5 and -qt6 ship a desktop file that depends on the icon
shipped by the -main package.
Take co-maintainership with ajacoutot@ (instead of pea@)

ok ajacoutot@

Not waiting the usual 7 days timeout so that this ships with
post-hackathon snapshot packages.