www/firefox-esr:  switch arm64 back to llvm 19
disable retguard via -fno-ret-protector, thanks robert@ for the hint.
apparently retguard on arm64 does something that xpcom doesnt like,
leading to the startup crashes seen and reported in #1973016.

this is not needed on amd64, where retguard stays enabled by default,
and -fno-ret-protector can be removed once arm64 is analyzed/fixed.