Stop attempting to overwrite the bootstrap libc.so
If the bootstrap was built using an old libc before the stdio opaque
FILE switch, then overwriting it with current libc is actively
detrimental. It made sense for the pinsyscalls transition, not for the
stdio transition. So fix up the "overwrite libc" bits (using tail -2
here wasn't really correct except on systems with a single libc.so
installed) and comment them for use the next time they prove useful.

Spotted with hppa's adastrap, which predates both the pinsyscall and
libc.so changes...

ok sthen@