switch back to the sqlite-src (vs sqlite-autoconf) distribution
disable -DSQLITE_DQS=0 when building the shell for now to reduce
unintentional behaviour difference from the previous version of this
port. it is enabled in upstream builds from sqlite-src since 3.41.0
(2023-02-21) but not currently in builds from sqlite-autoconf.

this relates to handling of string literals and trying to remove a
misfeature that was added early on for compatibility with mysql.
see https://sqlite.org/quirks.html#dblquote for more details, and
notes in patches/patch-main_mk in this commit.

i think we probably do want to change and follow upstream's default,
but let's let things settle like this for a bit first.

sleuthed by espie, this took a long time to find!