Add the follwoing local modifications to gcc to behave more like base-clang.
 - default to gnu17 (instead of gnu23)
 - turn -Wimplicit-function-declaration and -Wincompatible-pointer-types
   into simple warnings.

With this a large part of the ports tree can be compiled with gcc15 as
ports-gcc. Reverting the behaviour of -Wimplicit-function-declaration
and -Wincompatible-pointer-types is less then ideal but it fixing all the
bad code should be done on fast and popular clang archs first.

With and OK tb@, sthen@, jca@