Fix unwinding and C++ exceptions with gcc/11 and gcc/15
Upstream added an "&& defined(TARGET_DL_ITERATE_PHDR)" check that
specifically disabled the code walking PT_GNU_EH_FRAME, which resulted
in exceptions being broken. So explicitely list OpenBSD in the
more-or-less hardcoded list of platforms that need dl_iterate_phdr(3) in
gcc/configure.

This was spotted while building devel/boost on sparc64 but was
reproducible on amd64 and arm64. Hopefully that was the last hurdle
before we can attempt a move to lang/gcc/15 in the ports tree.
*crossing fingers*

Debugged with claudio@ and tb@, our discussions helped me a lot finding
the culprit. ok tb@ claudio@