merge textproc/libxml,-python (py3-libxml) into textproc/libxml; there's no direct equivalent any more (py3-lxml is the suggested replacement but it's not a drop-in) and having the old package lying around causes an issue for updates.
drop @pkgpath textproc/libxml; this was copied from PLIST-main but is the
default for a package built from ports/textproc/libxml without multipackages
anyway. update path is ok with just this; it was present directly in
PLIST-main so there's a valid update path. (that is ok because libxml
went from single- to multi-packages and back again; had it started as
multi-packages in the first place there would have been no @pkgpath in
PLIST-main - had that been the case, we'd need @pkgpath textproc,-main).

ok aja landry