Update Qt to 5.15.18 (+kde patchset collection)
The kde patchset collection has been rebased on top of qt 5.15.18.
I decided to switch all ports to use the KDE "kde/5.15" branch
(with a few exceptions where it does not exist because there are
no patches)

The good news is that QTWebEngine is now also at 5.15.18.
(based on chromium version: 87.0.4280.144, patched with security
patches up to: 130.0.6723.59)