Update to Python 3.13.11, ok kmos (maintainer), sthen
3.13.10 was a rather massive update and .11 adds a few fixes on top,
some of them for not too terrible security issues:

https://www.python.org/downloads/release/python-31310/
https://www.python.org/downloads/release/python-31311/

The annoying bit about this update is that our homegrown expat build
system does not install expat_config.h, which, strangely, is autoconf
results exposed in a public header (next to a few XML_ things). Until
that's sorted in base we get to patch the Python code since the expat
maintainer made the mitigation knobs only available behind some config
knobs, which the new Python code assumes to be enabled.