dxf2gcode: move to MODPY_PYBUILD, fix default paths for some optional tools (pdftops, pstoedit) while there.  ok tb@
rename the setuptools-compatible setup script so that it can be built as
a normal Python port via MODPY_PYBUILD; the previous method of running a
script to generate and install just the .egg file won't work with future
versions of setuptools.

issue with newer setuptools found by tb when bulk-testing.