json-glib: don't use --default-symver on ld.bfd archs, ok aja (maintainer)
It appears that --default-symver is broken in our ld.bfd. All binaries
in the sparc64 ${WRKBUILD} of json-glib contain broken .dynsym entries
such as

    57: dfdfdfdfdfdfdfdf 0xdfdfdfdfdfdfdfdf <processor specific>: 15 <processor
specific>: 13 PROTECTED [<other>: dc]  57311 <corrupt>

where __data_start should be. This isn't new. Release packages have this
as well in json-glib-{format,validate}.

What's new with meson 1.10.0 is that it tries to parse the .dynsym table:
https://github.com/mesonbuild/meson/pull/15106
for reproducibility reasons. This blows up due to this corrupted entry
(found by claudio and jca_p).

Longer term this wants a fix (probably in bfd/elflink.c) or workaround,
but for now let's unblock a good chunk of the ports tree broken with
meson 1.10.0.

ld.lld doesn't support this option, so nothing changes on most archs.
sthen informs me that nothing else uses --default-symver on sparc64.