remove "obsolete" quirks entry for glade, and merge devel/glade into lang/python/3 via @pkgpath + @conflict (no quirks entry required for that, because it is part of the update-set anyway).
glade in 7.8 depended on python 3.12 and in order that upgrades to
-current (and later 7.9) can work, we need to provide an upgrade path
that makes it disappear if installed. listing a port as obsolete does
not do this, it just leaves it installed, along with any packages
which it depends on (i.e. python 3.12 here), resulting in problems
if that causes a conflict.