import ports/astro/liberfa, ok tb
ERFA (Essential Routines for Fundamental Astronomy) is a C library
containing key algorithms for astronomy, and is based on the SOFA
library published by the International Astronomical Union (IAU),
https://www.iausofa.org/

ERFA is intended to replicate the functionality of SOFA (aside from
possible bugfixes in ERFA that have not yet been included in SOFA), but
is licensed under a three-clause BSD license to enable its compatibility
with a wide range of open source licenses. Permission for this release
has been obtained from the SOFA board, and is available in the LICENSE
file included in this source distribution.